Refactor some tests to fix inconsistencies in source structure
