+++ java/code/src/com/suse/manager/webui/controllers/test/LoginControllerTest.java
@@ -21,11 +21,11 @@ import com.redhat.rhn.manager.user.UserManager;
 import com.redhat.rhn.testing.RhnMockHttpServletRequest;
 import com.redhat.rhn.testing.RhnMockHttpServletResponse;
 import com.redhat.rhn.testing.RhnMockHttpSession;
+import com.redhat.rhn.testing.SparkTestUtils;
 import com.redhat.rhn.testing.UserTestUtils;
 
 import com.suse.manager.webui.controllers.login.LoginController;
 import com.suse.manager.webui.utils.LoginHelper;
-import com.suse.manager.webui.utils.SparkTestUtils;
 import com.suse.utils.Json;
 
 import java.io.UnsupportedEncodingException;
