+++ java/code/src/com/suse/manager/webui/controllers/test/VirtualHostManagerControllerTest.java
@@ -15,7 +15,7 @@
 
 package com.suse.manager.webui.controllers.test;
 
-import static com.suse.manager.webui.utils.SparkTestUtils.createMockRequestWithParams;
+import static com.redhat.rhn.testing.SparkTestUtils.createMockRequestWithParams;
 
 import com.redhat.rhn.common.hibernate.HibernateFactory;
 import com.redhat.rhn.domain.org.Org;
@@ -24,6 +24,7 @@ import com.redhat.rhn.domain.server.virtualhostmanager.VirtualHostManagerConfig;
 import com.redhat.rhn.domain.server.virtualhostmanager.VirtualHostManagerFactory;
 import com.redhat.rhn.testing.BaseTestCaseWithUser;
 import com.redhat.rhn.testing.RhnMockHttpServletResponse;
+import com.redhat.rhn.testing.SparkTestUtils;
 import com.redhat.rhn.testing.TestUtils;
 import com.redhat.rhn.testing.UserTestUtils;
 
@@ -31,7 +32,6 @@ import com.suse.manager.gatherer.GathererJsonIO;
 import com.suse.manager.gatherer.GathererRunner;
 import com.suse.manager.model.gatherer.GathererModule;
 import com.suse.manager.webui.controllers.VirtualHostManagerController;
-import com.suse.manager.webui.utils.SparkTestUtils;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
