+++ java/code/src/com/suse/manager/webui/utils/test/SparkTestUtilsTest.java
@@ -15,7 +15,7 @@
 
 package com.suse.manager.webui.utils.test;
 
-import com.suse.manager.webui.utils.SparkTestUtils;
+import com.redhat.rhn.testing.SparkTestUtils;
 
 import junit.framework.TestCase;
